Adding Gehler-Shi dataset support
Adding in the necessary scripts to evaluate on the Gehler-Shi dataset,
and fixing some other bugs that would prevent someone from evaluating on
any dataset. Also adding a processed version of the dataset, so that the
code works "out of the box", but with instructions on how to generate
that processed data from public sources.